Fixed pageCount

+++ api/krameriusplus-api/src/main/java/cz/inqool/dl4dh/krameriusplus/api/publication/object/PublicationDto.java
@@ -35,5 +35,5 @@ public abstract class PublicationDto extends DigitalObjectDto {
 
     private List<String> donator = new ArrayList<>();
 
-    private Long pageCount;
+    private Integer pageCount;
 }
